July 28th
Added some nice features to the menubar. First, when any of the options
open, the game pauses. When it closes, the game unpauses. The return to
menu option now has a confirmation dialogue. Also, the instructions
vary based on if it's a single or multiplayer game. Also, the
instructions and return to menu option only appear when in game.

Also changed from cardlayout to boxlayout.

Changed around the painting for singleplayer so that it would support
different sizes for a future version. Changed score to an integer. Also
changed the score display so that it would be right aligned.

In terms of multiplayer, used BitSet to track keys that were held down.
This allows support of multiple keys being held down. (If player 1 and
2 have the left key held down for instance.) Also added in a better
in-air collision detection system. Still needs work though.