July 30th - Resizing and statistics
Main menu has been updated with size radio buttons that allow the user
to choose the size of the window that will launch. Added in a
statistics option in the menu that displays how many of each piece has
fallen. Also simplified the action performed method in Tetris.java.
Also added in the resizing support for multiplayer. Still need to fix
multiplayer’s issue when two keys are pressed together one of them is
double pressed.